% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect Function}
\usage{
connect_db(server, db, uid, pwd)
}
\arguments{
\item{server}{Database Server Url/IP Address}

\item{db}{Database Name}

\item{uid}{User id}

\item{pwd}{Password}
}
\description{
This function allows you to connect to a MS SqlServer Database
}
\examples{
dbHandle <- connect_db("server.com","db_test","admin","password")
}
\keyword{Connect}
