% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_entity_gen.R
\name{connect_entity_gen}
\alias{connect_entity_gen}
\title{Connect to Database Server and Generate Entity Details Function}
\usage{
connect_entity_gen(server, db, uid, pwd)
}
\arguments{
\item{server}{Database Server Url/IP Address}

\item{db}{Database Name}

\item{uid}{User id}

\item{pwd}{Password}
}
\description{
This function connects to database and creates entity details
}
\examples{
dbHandle <- connect_entity_gen("server.com","db_test","admin","password")
}
\keyword{Connect}
\keyword{Entity}
\keyword{Generation}
