% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{exec}
\alias{exec}
\title{Executes the query}
\usage{
exec(query, dbHandle)
}
\arguments{
\item{query}{full query for execution}

\item{dbHandle}{Database Handle to use for query}
}
\description{
This function executes the query
}
\examples{
df <- exec("",NULL)
}
\keyword{exec}
