% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R, R/where.R
\name{from}
\alias{from}
\alias{from}
\title{Generates from part for the query}
\usage{
from(query = NULL, ...)

from(query = NULL, ...)
}
\arguments{
\item{query}{Existing query}

\item{...}{where list}

\item{data}{Table or View to query from}

\item{query}{Existing query}
}
\description{
This function creates a from part for the query

This function generates "where" part of a query
}
\examples{
from_tb <- from(dbHandle$tabl_dbo.table)
from_tb <- where("", dbHandle$col_tabl_dbo.col1, "=", "test1",dbHandle$col_tabl_dbo.col1, >, val1)
}
\keyword{from}
\keyword{where}
