% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trims leading and trailing whitespaces}
\usage{
trim(x)
}
\arguments{
\item{x}{string to trim}
}
\description{
This function trims leading and trailing whitespaces
}
\examples{
df <- trim(" text containing leading and trailing whitespaces. ")
#text containing leading and trailing whitespaces.
}
\keyword{leading}
\keyword{trailing}
\keyword{trim}
\keyword{whitespaces}
