% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{from}
\alias{from}
\title{Generates from part fo the query}
\usage{
from(data)
}
\arguments{
\item{data}{Table or View to query from}
}
\description{
This function creates a from part fo the query
}
\examples{
from_tb <- from(dbHandle$tabl_dbo.table)
}
\keyword{from}
