% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{from}
\alias{from}
\title{Connect to Database Server and Generate Entity Details Function}
\usage{
from(data)
}
\arguments{
\item{data}{Table or View to query from}
}
\description{
This function connects to database and creates entity details
}
\examples{
from_tb <- from(dbHandle$tabl_dbo.table)
}
\keyword{from}
