% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{from}
\alias{from}
\title{Generates from part for the query}
\usage{
from(query = NULL, data)
}
\arguments{
\item{query}{Existing query}

\item{data}{Table or View to query from}
}
\description{
This function creates a from part for the query
}
\examples{
from_tb <- from("","tableName")
}
\keyword{from}
