% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Generates Select part fo the query}
\usage{
select(query = NULL, ..., top = NULL, distinct = NULL)
}
\arguments{
\item{query}{Existing query}

\item{...}{Fields to query}

\item{top}{Top n rows of the data}

\item{distinct}{True or FALSE}
}
\description{
This function creates "Select" part fo the query
}
\examples{

#Returns Top 100 rows
from_tb <- select("","COl1","COl2","Col3", top=100)

#Returns distinct rows
from_tb <- select("","COl1","COl2","Col3", distinct=TRUE)
}
\keyword{from}
