% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Generates where part of a query}
\usage{
where(query = NULL, ...)
}
\arguments{
\item{query}{Existing query}

\item{...}{where list}
}
\description{
This function generates "where" part of a query
}
\examples{
where_cond <- where("", "col1", "=", "test1", "AND", "col2", ">", 25)
}
\keyword{where}
